Move functions to separate file, handle globals
