+++ index.js
@@ -1,8 +1,8 @@
 const fs = require('fs')
 const { performance } = require('perf_hooks')
 const globby = require('globby')
-const { sha256 } = require('js-sha256')
-const { JSDOM } = require('jsdom')
+
+const { mergeWithDefaultPolicies, createFileProcessor, buildCSPArray, splitToGlobalAndLocal } = require('./functions.js')
 
 module.exports = {
   onPostBuild: async ({ inputs }) => {
@@ -19,96 +19,24 @@ module.exports = {
     const paths = await globby(lookup)
     console.info(`Found ${paths.length} HTML ${paths.length === 1 ? 'file' : 'files'}`)
 
-    const processFile = createFileProcessor(buildDir, mergedPolicies, disablePolicies, disableGeneratedPolicies)
+    const processFile = createFileProcessor(buildDir, disableGeneratedPolicies)
 
     const processedFileHeaders = await Promise.all(
       paths.map(path => fs.promises.readFile(path, 'utf-8').then(processFile(path)))
     )
 
-    const file = processedFileHeaders
-      .map(({ webPath, cspString }) => `${webPath}\n  Content-Security-Policy: ${cspString}`)
-      .join('\n')
+    const { globalHeaders, localHeaders } = processedFileHeaders
+      .reduce(splitToGlobalAndLocal, { globalHeaders: [], localHeaders: [] })
+
+    const file = globalHeaders.concat(...localHeaders)
+      .map(({ webPath, cspObject }) => {
+        const cspString = buildCSPArray(mergedPolicies, disablePolicies, cspObject).join(' ')
+        return `${webPath}\n  Content-Security-Policy: ${cspString}`
+      }).join('\n')
 
     fs.appendFileSync(`${buildDir}/_headers`, file)
 
     const completedTime = performance.now() - startTime
     console.info(`Saved at ${buildDir}/_headers - ${(completedTime / 1000).toFixed(2)} seconds`)
   },
-}
-
-function mergeWithDefaultPolicies (policies) {
-  const defaultPolicies = {
-    defaultSrc: '',
-    childSrc: '',
-    connectSrc: '',
-    fontSrc: '',
-    frameSrc: '',
-    imgSrc: '',
-    manifestSrc: '',
-    mediaSrc: '',
-    objectSrc: '',
-    prefetchSrc: '',
-    scriptSrc: '',
-    scriptSrcElem: '',
-    scriptSrcAttr: '',
-    styleSrc: '',
-    styleSrcElem: '',
-    styleSrcAttr: '',
-    workerSrc: '',
-    baseUri: '',
-    formAction: '',
-    frameAncestors: '',
-  }
-
-  return {...defaultPolicies, ...policies}
-}
-
-function createFileProcessor (buildDir, mergedPolicies, disablePolicies, disableGeneratedPolicies) {
-  return path => file => {
-    const dom = new JSDOM(file)
-    const shouldGenerate = (key) => !(disableGeneratedPolicies || []).includes(key)
-    const generateHashesFromElement = generateHashes(dom, element => element.innerHTML)
-    const generateHashesFromStyle = generateHashes(dom, element => element.getAttribute('style'))
-
-    const scripts = shouldGenerate('scriptSrc') ? generateHashesFromElement('script') : []
-    const styles = shouldGenerate('styleSrc') ? generateHashesFromElement('style') : []
-    const inlineStyles = shouldGenerate('styleSrc') ? generateHashesFromStyle('[style]') : []
-
-    const webPath = path.replace(new RegExp(`^${buildDir}(.*)index\\.html$`), '$1')
-    const cspString = buildCSPArray(mergedPolicies, disablePolicies, {
-      scriptSrc: scripts,
-      styleSrc: [...styles, ...inlineStyles],
-    }).join(' ')
-
-    return {
-      webPath,
-      cspString,
-    }
-  }
-}
-
-function generateHashes (dom, getPropertyValue) {
-  return selector => {
-    const hashes = new Set()
-    for (const matchedElement of dom.window.document.querySelectorAll(selector)) {
-      const value = getPropertyValue(matchedElement)
-      if (value.length) {
-        const hash = sha256.arrayBuffer(value)
-        const base64hash = Buffer.from(hash).toString('base64')
-        hashes.add(`'sha256-${base64hash}'`)
-      }
-    }
-    return Array.from(hashes)
-  }
-}
-
-function buildCSPArray (allPolicies, disablePolicies, hashes) {
-  const camelCaseToKebabCase = (string) => string.replace(/([a-z0-9])([A-Z])/g, '$1-$2').toLowerCase()
-
-  return Object.entries(allPolicies)
-    .filter(([key, defaultPolicy]) => (hashes[key] || defaultPolicy) && !(disablePolicies || []).includes(key))
-    .map(([key, defaultPolicy]) => {
-      const policy = `${hashes[key] && hashes[key].join(' ') || ''} ${defaultPolicy}`;
-      return `${camelCaseToKebabCase(key)} ${policy.trim()};`
-    })
 }
